feat(payment): add payment feature

+++ src/main/java/vn/id/nonglam/kltn/kltn/dto/response/search/SearchHotelResponse.java
@@ -1,8 +1,9 @@
 package vn.id.nonglam.kltn.kltn.dto.response.search;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record SearchHotelResponse(UUID id, String name, String thumbnail, String street, String ward,
                                   String province, double latitude, double longitude,
                                   String description, int viewCount, double avgRating, int totalComment,
-                                  double minPrice, double maxPrice) {}
+                                  BigDecimal minPrice, BigDecimal maxPrice) {}
